requirements.txt
Django>=5.0,<6
python-dotenv>=1.0.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg[binary]>=3.2
PyJWT>=2.8
httpx>=0.27
pytest>=8.0
pytest-django>=4.8
pytest-asyncio>=0.24
testcontainers>=4.7
groq>=0.4.0
